♻️ refactor ResultPage
